Move get_sector_safeguard_facilities into safeguard module
This was created for the oil and gas sector, but is helpful for other sectors which want to implement Safeguard Mechanism integration.
